+++ facundo/02-oop-inheritance/Actor.js
@@ -1,0 +1,7 @@
+class Actor {
+    constructor(name, age) {
+        this.name = name;
+        this.age = age;
+    }
+}
+module.exports=Actor;
