+++ facundo/02-oop-inheritance/Logger.js
@@ -1,0 +1,13 @@
+
+/**
+ * 
+ * 
+ * @class Logger
+ * Class for exercise 6 it just log something in the console
+ */
+class Logger {
+    log(info) {
+        console.log(info);
+    }
+}
+module.exports=Logger;
