+++ facundo/02-oop-inheritance/index.html
@@ -1,5 +1,5 @@
 <head >
-    <script src="script.js" type="text/javascript"></script>
+    <script src="bundle.js" type="text/javascript"></script>
 </head>
 <body>
