+++ facundo/02-oop-inheritance/package.json
@@ -1,0 +1,25 @@
+{
+  "name": "02-oop-inheritance",
+  "version": "1.0.0",
+  "main": "script.js",
+  "scripts": {
+    "start": "index.js",
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "author": "",
+  "license": "ISC",
+  "dependencies": {
+    "babel-preset-env": "^1.3.3",
+    "babelify": "^6.4.0",
+    "browserify": "^14.3.0",
+    "gulp": "^3.9.1",
+    "gulp-util": "^3.0.8",
+    "vinyl-source-stream": "^1.1.0"
+  },
+  "devDependencies": {
+    "babel-cli": "^6.24.1",
+    "babel-preset-env": "^1.3.3",
+    "gulp": "^3.9.1"
+  },
+  "description": ""
+}
